Adds support to skip server SSL cert validation
Currently, running with SSL URIs, certificate was always checked to be valid.
This was troublesome in development, where the API server certificate is self-
signed. This patch adds a new configurable constant in options.rb (SSL_VERIFY).
This constant can be set to·OpenSSL::SSL::VERIFY_PEER or
OpenSSL::SSL::VERIFY_NONE respectiveley to check, not check the server
certificate during API calls. The default is TO CHECK the certificate.